Modular pdf file/OCR recognition engine
