+++ src/pdfdeal/Watch/config.py
@@ -1,0 +1,3 @@
+import os
+def init_config():
+    from ..FileTools.ocr import BUILD_IN_OCR
